get rid of compiler warnings (#250) ci run

+++ DearPyGui/src/Core/AppItems/mvBasicItems.h
@@ -302,7 +302,7 @@ namespace Marvel {
 			auto conflictingflagop = [dict](const std::vector<std::string>& keywords, std::vector<int> flags, int& mflags)
 			{
 
-				for (int i = 0; i < keywords.size(); i++)
+				for (size_t i = 0; i < keywords.size(); i++)
 				{
 					if (PyObject* item = PyDict_GetItemString(dict, keywords[i].c_str()))
 					{
